Fix crash when main branch is not found
If the main branch of a repository is not named 'main', 'master',
or 'trunk' or does not exist locally, the 'clip' command crashes:

```
❯ clip
panic: runtime error: invalid memory address or nil pointer dereference
[signal SIGSEGV: segmentation violation code=0x1 addr=0x20 pc=0x10b99df]

goroutine 1 [running]:
main.main()
	/usr/local/bin/cmd/clip/clip.go:102 +0x51f
```

This occurs because the code assumes that the master branch has been
found and attempts to dereference `details["_trunk_"]`, which does
not exist.

This fix throws an error if the main branch is not found locally.